refactor: move survey-card component to common
