+++ app/profile/page.tsx
@@ -1,5 +1,5 @@
 import { ProfileHeader } from '~/components/profile/profile-header'
-import { SurveyCard } from '~/components/profile/profile-survey-card'
+import { SurveyCard } from '~/components/common/survey-card'
 import { SiteHeader } from '~/components/common/site-header'
 
 function ProfilePage() {
