+++ src/components/common/survey-card.tsx
